impl(storage): move v1 module to its own file
